fixed bugs, got rid of useless components
