+++ src/components/EditProfile.js
@@ -1,9 +1,40 @@
 import React, { useState } from "react";
 import axios from "axios";
-import { Button, Checkbox, Form } from "semantic-ui-react";
+// import { Button, Checkbox } from "semantic-ui-react";
+import { device } from "../styles/breakpoints";
 import styled from "styled-components";
+import { makeStyles } from "@material-ui/core/styles";
+import MenuItem from "@material-ui/core/MenuItem";
+import Button from "@material-ui/core/Button";
+import TextField from "@material-ui/core/TextField";
+
+const useStyles = makeStyles(theme => ({
+  container: {
+    display: "flex",
+    flexWrap: "wrap"
+  },
+  textField: {
+    width: "70%",
+    maxWidth: "600px",
+    marginLeft: theme.spacing(1),
+    marginRight: theme.spacing(1)
+  },
+  dense: {
+    marginTop: theme.spacing(2)
+  },
+  menu: {
+    width: 200
+  },
+  button: {
+    margin: theme.spacing(1),
+    // marginTop: "3rem",
+    width: "40%",
+    maxWidth: "400px"
+  }
+}));
 
 function EditProfile(props) {
+  const classes = useStyles();
   const [input, setInput] = useState({
     email: props.currentUser.email,
     picture: props.currentUser.picture,
@@ -50,67 +81,98 @@ function EditProfile(props) {
   };
 
   return (
-    <div>
-      <div>
-        <h1>Edit</h1>
+    <FormContainer>
+      <h1>Edit Profile</h1>
 
-        <Form onSubmit={updateHandler} className="formstyle">
-          <Form.Field>
-            <label>Username</label>
-            <input
-              type="text"
-              value={input.username}
-              name="username"
-              onChange={handleInput}
-              placeholder="Username"
-            />
-          </Form.Field>
+      <Form onSubmit={updateHandler}>
+        <TextField
+          id="outlined-with-placeholder"
+          label="Username"
+          placeholder="Profile Picture"
+          className={classes.textField}
+          margin="normal"
+          variant="outlined"
+          name="username"
+          value={input.username}
+          onChange={handleInput}
+        />
 
-          <Form.Field>
-            <label>Profile Picture</label>
-            <input
-              type="text"
-              value={input.picture}
-              name="picture"
-              onChange={handleInput}
-              placeholder="Profile Picture"
-            />
-          </Form.Field>
+        <TextField
+          id="outlined-with-placeholder"
+          label="Picture"
+          placeholder="Profile Picture"
+          className={classes.textField}
+          margin="normal"
+          variant="outlined"
+          name="picture"
+          value={input.picture}
+          onChange={handleInput}
+        />
 
-          <Form.Field>
-            <label>Email</label>
-            <input
-              type="text"
-              value={input.email}
-              name="email"
-              onChange={handleInput}
-              placeholder="Email"
-            />
-          </Form.Field>
+        <TextField
+          id="outlined-with-placeholder"
+          label="Email"
+          placeholder="Profile Picture"
+          className={classes.textField}
+          margin="normal"
+          variant="outlined"
+          name="email"
+          value={input.email}
+          onChange={handleInput}
+        />
 
-          <Form.Field>
-            <label>Zip Code</label>
-            <input
-              type="text"
-              value={input.zipCode}
-              name="zip code"
-              onChange={handleInput}
-              placeholder="Zip Code"
-            />
-          </Form.Field>
-          <Button type="submit">Submit</Button>
-          <Button onClick={props.handleEdit}>Back</Button>
-        </Form>
-        {/* <button onClick={props.handleEdit}>Back</button> */}
-      </div>
-    </div>
+        <TextField
+          id="outlined-with-placeholder"
+          label="Zip Code"
+          placeholder="Profile Picture"
+          className={classes.textField}
+          margin="normal"
+          variant="outlined"
+          name="zipCode"
+          value={input.zipCode}
+          onChange={handleInput}
+        />
+
+        <Button
+          type="submit"
+          variant="outlined"
+          color="primary"
+          className={classes.button}
+        >
+          Submit
+        </Button>
+
+        <Button
+          variant="outlined"
+          color="secondary"
+          onClick={props.handleEdit}
+          className={classes.button}
+        >
+          Back
+        </Button>
+      </Form>
+
+      {/* <button type="submit">Submit</button>
+      <button onClick={props.handleEdit}>Back</button> */}
+    </FormContainer>
   );
 }
 
-const Container = styled.div`
+const FormContainer = styled.div`
+  margin-top: 3%;
   display: flex;
   flex-direction: column;
   justify-content: center;
+  h1 {
+    text-align: center;
+  }
+`;
+
+const Form = styled.form`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 `;
 
 export default EditProfile;
