+++ src/components/Profile.js
@@ -2,7 +2,6 @@ import React from "react";
 import axios from "axios";
 import { useState, useEffect } from "react";
 import EditProfile from "./EditProfile";
-import EditPro from "./EditPro";
 import { Button, Card, Icon, Image, Table } from "semantic-ui-react";
 import EditIssue from "./EditIssue";
 import styled from "styled-components";
@@ -12,16 +11,19 @@ import EditIcon from "@material-ui/icons/Edit";
 import RoomIcon from "@material-ui/icons/Room";
 import IssuesList from "./IssuesList";
 import { device } from "../styles/breakpoints";
+import CircularProgress from "@material-ui/core/CircularProgress";
 
 function Profile(props) {
   const [currentUser, setCurrentUser] = useState("");
   const [isEditingUser, setIsEditingUser] = useState(false);
   const [isEditingIssue, setIsEditingIssue] = useState(false);
   const [issueToUpdate, setIssueToUpdate] = useState({});
+  const [loading, setLoading] = useState(false);
   let token = JSON.parse(localStorage.getItem("token"));
   let localId = JSON.parse(localStorage.getItem("id"));
 
   useEffect(() => {
+    setLoading(true);
     axios
       .get(`https://co-make.herokuapp.com/users/${localId}/issues`, {
         headers: {
@@ -31,6 +33,7 @@ function Profile(props) {
       .then(res => {
         console.log("USER DATA FROM SERVER", res);
         setCurrentUser(res.data);
+        setLoading(false);
       })
       .catch(err => console.log("OH NO AN ERROR HAPPENED", err));
   }, []);
@@ -38,6 +41,7 @@ function Profile(props) {
   const handleEdit = e => {
     setIsEditingUser(!isEditingUser);
   };
+
   const handleEditIssue = id => {
     let thisIssue = currentUser.issues.filter(issue => issue.id === id);
     setIssueToUpdate(...thisIssue);
@@ -53,125 +57,84 @@ function Profile(props) {
       })
       .then(res => {
         axios
-           .get(`https://co-make.herokuapp.com/users/${localId}/issues`, {
-              headers: {
-                Authorization: token
-              }
-             })
-            .then( res => {
-            console.log("USER DATA FROM SERVER", res)
-            setCurrentUser(res.data)
-            setLoading(false)
+          .get(`https://co-make.herokuapp.com/users/${localId}/issues`, {
+            headers: {
+              Authorization: token
+            }
           })
-            .catch( err => console.log("OH NO AN ERROR HAPPENED", err))
-        },[])
-
-        const handleEdit = e => {
-          setIsEditingUser(!isEditingUser);
-        };
-        const handleEditIssue = id => {
-          let thisIssue = currentUser.issues.filter( issue => issue.id === id);
-          setIssueToUpdate(...thisIssue)
-          setIsEditingIssue(!isEditingIssue)
-        }
-
-        const deleteIssue = id => {
-          axios
-            .delete(`https://co-make.herokuapp.com/issues/${id}`, {
-              headers: {
-                Authorization: token
-              }
-            })
-            .then( res => {
-
-              axios.get(`https://co-make.herokuapp.com/users/${localId}/issues`, {
-                headers: {
-                  Authorization: token
-                }
-               }).then( res => {
-                 console.log("NEW DATA FROM SERVER", res)
-                 setCurrentUser(res.data)
-               }).catch( err => {
-                 console.log("OH NO", err)
-               })
-            })
-            .catch( err => {
-              console.log("Error on delete", err)
-            })
-          }
+          .then(res => {
+            console.log("NEW DATA FROM SERVER", res);
+            setCurrentUser(res.data);
+          })
+          .catch(err => {
+            console.log("OH NO", err);
+          });
+      })
+      .catch(err => {
+        console.log("Error on delete", err);
+      });
+  };
 
   return (
-    <Container>
-      {isEditingUser ? (
-        <EditPro
-          handleEdit={handleEdit}
-          currentUser={currentUser}
-          setCurrentUser={setCurrentUser}
-        />
-      ) : isEditingIssue ? (
-        <EditIssue
-          setCurrentUser={setCurrentUser}
-          setIsEditingIssue={setIsEditingIssue}
-          isEditingIssue={isEditingIssue}
-          issueToUpdate={issueToUpdate}
-        />
+    <>
+      {loading ? (
+        <LoaderWrapper>
+          <CircularProgress />
+        </LoaderWrapper>
       ) : (
-        <ProfileContainer>
-          <Banner>
-            <BannerImage src={skyline} alt="Banner" />
-          </Banner>
-          <ImageCrop>
-            <ProfileImage
-              src={currentUser.picture ? currentUser.picture : avi}
-              alt="Profile"
+        <Container>
+          {isEditingUser ? (
+            <EditProfile
+              handleEdit={handleEdit}
+              currentUser={currentUser}
+              setCurrentUser={setCurrentUser}
             />
-          </ImageCrop>
+          ) : isEditingIssue ? (
+            <EditIssue
+              setCurrentUser={setCurrentUser}
+              setIsEditingIssue={setIsEditingIssue}
+              isEditingIssue={isEditingIssue}
+              issueToUpdate={issueToUpdate}
+            />
+          ) : (
+            <ProfileContainer>
+              <Banner>
+                <BannerImage src={skyline} alt="Banner" />
+              </Banner>
+              <ImageCrop>
+                <ProfileImage
+                  src={currentUser.picture ? currentUser.picture : avi}
+                  alt="Profile"
+                />
+              </ImageCrop>
 
-          <MainProfile>
-            <UserInfo>
-              <h1>{currentUser.username}</h1>
-              <h4>
-                <RoomIcon />
-                {currentUser.zipCode}
-              </h4>
-              <h4>{currentUser.email}</h4>
-              <h6 onClick={handleEdit}>
-                <EditIcon /> Edit Profile
-              </h6>
-            </UserInfo>
+              <MainProfile>
+                <UserInfo>
+                  <h1>{currentUser.username}</h1>
+                  <h4>
+                    <RoomIcon />
+                    {currentUser.zipCode}
+                  </h4>
+                  <h4>{currentUser.email}</h4>
+                  <h6 onClick={handleEdit}>
+                    <EditIcon /> Edit Profile
+                  </h6>
+                </UserInfo>
 
-            <UsersIssues>
-              <h3>
-                {currentUser.username} has voted on # issues in the community
-              </h3>
-              <h3>{currentUser.username}'s created issues:</h3>
-              <IssuesList issues={currentUser.issues} />
-            </UsersIssues>
-          </MainProfile>
-          <footer className="footer-nav">
-            <Nav className="bottom-nav">
-              <Button.Group widths="3" size="big">
-                <Button
-                  icon="list alternate outline"
-                  content="Feed"
-                  onClick={() => props.history.push("/")}
-                />
-                <Button
-                  icon="add"
-                  content="Create Issue"
-                  onClick={() => props.history.push("/addIssue")}
-                />
-                <Button
-                  icon="user"
-                  content="Profile"
-                  onClick={() => props.history.push(`/profile/${localId}`)}
-                />
-              </Button.Group>
-            </Nav>
-          </footer>
-        </ProfileContainer>
+                <UsersIssues>
+                  <h3>
+                    {currentUser.username} has voted on # issues in the
+                    community
+                  </h3>
+                  <h3>{currentUser.username}'s created issues:</h3>
+                  <IssuesList issues={currentUser.issues} />
+                </UsersIssues>
+              </MainProfile>
+            </ProfileContainer>
+          )}
+        </Container>
       )}
-    </Container>
+    </>
   );
 }
 
@@ -186,7 +149,6 @@ const ProfileContainer = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
-  margin-top: 7.1%;
   padding-bottom: 5rem;
   width: 100%;
   background-color: #f8f8ff;
@@ -199,7 +161,7 @@ const BannerImage = styled.img`
   border-bottom: 10px solid black;
   max-height: 225px;
   object-fit: cover;
-  `
+`;
 
 const ImageCrop = styled.div`
   width: 150px;
@@ -245,38 +207,21 @@ const UserInfo = styled.div`
   h4 {
     font-size: 2rem;
   }
-`
+`;
 
-  const LoaderWrapper = styled.div`
+const LoaderWrapper = styled.div`
   width: 100%;
-  display:flex;
-  justify-content:center;
+  display: flex;
+  justify-content: center;
   align-items: center;
-  height: 50vh
-  `
-
+  height: 50vh;
+`;
 
 const UsersIssues = styled.div`
-  margin-left: 2rem;
   text-align: center;
   @media ${device.tablet} {
     margin-top: 2rem;
   }
-`;
-
-const Nav = styled.nav`
-  display: flex;
-  border: none;
-  justify-content: space-evenly;
-  align-items: center;
-  font-family: "helvetica", sans serif;
-  a {
-    color: #eb472c;
-    text-decoration: none;
-  }
-  height: 50px;
-  font-size: 1.2rem;
-  font-weight: bold;
 `;
 
 export default Profile;
